+++ Flow/src/utopia/flow/view/mutable/eventful/LockablePointer.scala
@@ -2,31 +2,14 @@ package utopia.flow.view.mutable.eventful
 
 import utopia.flow.util.TryExtensions._
 import utopia.flow.util.logging.Logger
-import utopia.flow.view.mutable.PointerFactory
+import utopia.flow.view.mutable.LoggingPointerFactory
 import utopia.flow.view.template.eventful.{AbstractMayStopChanging, Changing, ChangingWrapper}
 
-import scala.annotation.unused
-import scala.language.implicitConversions
 import scala.util.Try
 
-object LockablePointer
+object LockablePointer extends LoggingPointerFactory[LockablePointer]
 {
-	// COMPUTED --------------------------
-	
-	/**
-	  * @param log Logging implementation for handling failures in change-event -handling
-	  * @return Factory for constructing lockable pointers
-	  */
-	def factory(implicit log: Logger): PointerFactory[LockablePointer] = new LockablePointerFactory()
-	
-	
-	// IMPLICIT --------------------------
-	
-	implicit def objectToFactory(@unused o: LockablePointer.type)
-	                            (implicit log: Logger): PointerFactory[LockablePointer] = factory
-	
-	
-	// OTHER    --------------------------
+	// IMPLEMENTED    ----------------------
 	
 	/**
 	  * @param initialValue Initial value to assign to this pointer
@@ -34,14 +17,38 @@ object LockablePointer
 	  * @tparam A Type of values held within this pointer
 	  * @return A new pointer
 	  */
-	def apply[A](initialValue: A)(implicit log: Logger) = new LockablePointer[A](initialValue)
+	override def apply[A](initialValue: A)(implicit log: Logger): LockablePointer[A] =
+		new _LockablePointer[A](initialValue)
 	
 	
 	// NESTED   ---------------------------
 	
-	private class LockablePointerFactory(implicit log: Logger) extends PointerFactory[LockablePointer]
+	private class _LockablePointer[A](initialValue: A)(implicit log: Logger)
+		extends AbstractMayStopChanging[A] with LockablePointer[A]
 	{
-		override def apply[A](initialValue: A): LockablePointer[A] = LockablePointer(initialValue)
+		// ATTRIBUTES   -------------------------
+		
+		private var _value = initialValue
+		private var _locked = false
+		
+		override lazy val readOnly: Changing[A] = ChangingWrapper(this)
+		
+		
+		// IMPLEMENTED  -------------------------
+		
+		override def value: A = _value
+		override def locked = _locked
+		
+		override def lock() = {
+			_locked = true
+			declareChangingStopped()
+		}
+		
+		override protected def assignToUnlocked(newValue: A): Unit = {
+			val oldValue = _value
+			_value = newValue
+			fireEventIfNecessary(oldValue, newValue).foreach { effect => Try { effect() }.log }
+		}
 	}
 }
 
@@ -51,34 +58,26 @@ object LockablePointer
   * @author Mikko Hilpinen
   * @since 26.7.2023, v2.2
   */
-class LockablePointer[A](initialValue: A)(implicit log: Logger)
-	extends AbstractMayStopChanging[A] with Lockable[A] with EventfulPointer[A]
+trait LockablePointer[A] extends Lockable[A] with EventfulPointer[A]
 {
-	// ATTRIBUTES   -------------------------
+	// ABSTRACT -----------------------------
 	
-	private var _value = initialValue
-	private var _locked = false
-	
-	override lazy val readOnly: Changing[A] = ChangingWrapper(this)
+	/**
+	 * Assigns a new value to this pointer, also firing change events.
+	 * This method is only called for non-locked pointers.
+	 * @param newValue New value to assign to this pointer
+	 */
+	protected def assignToUnlocked(newValue: A): Unit
 	
 	
 	// IMPLEMENTED  -------------------------
 	
-	override def value: A = _value
 	@throws[IllegalStateException]("If this pointer has been locked")
 	override def value_=(newValue: A): Unit = {
-		if (_locked)
+		if (locked)
 			throw new IllegalStateException("This pointer has been locked")
 		else
-			_set(newValue)
-	}
-	
-	override def locked = _locked
-	
-	override def lock() = {
-		_locked = true
-		// Discards all listeners, since they won't be informed about anything anymore
-		declareChangingStopped()
+			assignToUnlocked(newValue)
 	}
 	
 	
@@ -93,17 +92,10 @@ class LockablePointer[A](initialValue: A)(implicit log: Logger)
 	  */
 	def trySet(value: A) = {
 		if (locked)
-			value == _value
+			value == this.value
 		else {
-			_set(value)
+			assignToUnlocked(value)
 			true
 		}
-	}
-	
-	// Expects locking status to be checked
-	private def _set(newValue: A) = {
-		val oldValue = _value
-		_value = newValue
-		fireEventIfNecessary(oldValue, newValue).foreach { effect => Try { effect() }.log }
 	}
 }